pyproject.toml
[tool.poetry]
name = "proto-services"
version = "0.1.0"
description = ""
authors = ["Maxim Nalimov <nalimov-maxim@rambler.ru>"]
readme = "README.md"
packages = [{include = "proto_services"}]

[tool.poetry.dependencies]
python = "^3.9.16"
grpcio-tools = "^1.33.2"
grpcio = "^1.51.3"
protobuf = "^3.14.0"
pytest = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
